+++ plone/app/controlpanel/tests/__init__.py
@@ -1,3 +1,1 @@
-# Other packages may find this useful
-
-from plone.app.controlpanel.tests.cptc import ControlPanelTestCase
+#